Moves insert and update methods to entities file, and made imports conditional.
